pyproject.toml
[tool.poetry]
name = "todos-drf-api"
version = "0.1.0"
description = ""
authors = ["Lobo <matheuscardoso1@poli.ufrj.br>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
Markdown = "^3.3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
